+++ script/tool/lib/src/drive_examples_command.dart
@@ -43,6 +43,10 @@ class DriveExamplesCommand extends PackageLoopingCommand {
       help:
           'Runs the driver tests in Dart VM with the given experiments enabled.',
     );
+    argParser.addFlag(
+      kEnableSoftwareRendering,
+      help: 'Renders the driver tests using the Skia software backend.',
+    );
   }
 
   @override
@@ -337,6 +341,8 @@ class DriveExamplesCommand extends PackageLoopingCommand {
             ...deviceFlags,
             if (enableExperiment.isNotEmpty)
               '--enable-experiment=$enableExperiment',
+            if (getBoolArg(kEnableSoftwareRendering))
+              '--enable-software-rendering',
             '--driver',
             getRelativePosixPath(driver, from: example.directory),
             '--target',